chore(main): rename entry file
