+++ package.json
@@ -2,18 +2,18 @@
   "name": "typescript-starter",
   "version": "1.1.2",
   "description": "Run TypeScript programs with eslint, prettier and semantic release support.",
-  "main": "build/app.js",
+  "main": "build/main.js",
   "author": "Fenando López Martínez",
   "license": "MIT",
   "scripts": {
     "build": "tsc",
     "commit": "git-cz",
-    "dev": "ts-node-dev --respawn --no-notify --transpileOnly src/app.ts",
-    "dev:inspect": "ts-node-dev --inspect-brk --respawn --no-notify --transpileOnly src/app.ts",
+    "dev": "ts-node-dev --respawn --no-notify --transpileOnly src/main.ts",
+    "dev:inspect": "ts-node-dev --inspect-brk --respawn --no-notify --transpileOnly src/main.ts",
     "lint": "eslint . --ext .js,.ts",
     "prebuild": "rm -rf ./build",
     "semantic-release": "semantic-release",
-    "start": "node build/app.js",
+    "start": "node build/main.js",
     "test": "jest --passWithNoTests --runInBand",
     "test:watch": "jest --passWithNoTests --runInBand --watch",
     "test:coverage": "jest --runInBand --coverage"
