+++ src/__tests__/main.unit.test.ts
@@ -1,4 +1,4 @@
-import { greetingUppercase } from './../app';
+import { greetingUppercase } from './../main';
 
 test('formats the greeting correctly', () => {
   expect(greetingUppercase('world')).toMatchSnapshot();
